[C8] Split chamfer.go by responsibility (#2232)
model/feature/chamfer.go was 612 LOC, over the 500-line ceiling. Split the
chamfer feature and its corner tool; each file is now under the ceiling:

- chamfer.go (357): the chamfer feature entry, wedge-cut driver, setback/mode
  resolution, edge resolution and the edge corner frame.
- chamfer_corner.go (210): the flat three-edge-corner blend — the tetra cut
  filling the notch where three selected edges meet, solved from the three
  edges' chamfer planes' triple intersection.
- chamfer_tetra.go (76): the tetra tool body builder (six edges keyed by
  ascending vertex-index pair, four faces sharing them consistently).

Behaviour-preserving: split by exact line ranges, verified every non-comment
code line is byte-identical (only per-file import redistribution differs);
imports resolved per file with goimports. Chamfer corpus + archguard pass
before and after with no golden updates. Net delta of recognizers / tolerance
constants / fallback sites / type assertions: 0.

Note: the tetra builder is also flagged C6-a (guarded edge reuse); that
guarding is the separate phase-2 task.

Part of #2183 (C8: structure/size limits). Closes #2232.

(cherry picked from commit c629e71142cbb34ac80e774207ca2d1e5338961e)
